fix: handle clickhouse optionally
