+++ controlplane/src/core/bufservices/cache-warmer/computeCacheWarmerOperations.ts
@@ -30,20 +30,8 @@ export function computeCacheWarmerOperations(
     req.namespace = req.namespace || DefaultNamespace;
 
     const fedGraphRepo = new FederatedGraphRepository(logger, opts.db, authContext.organizationId);
-    const subgraphRepo = new SubgraphRepository(logger, opts.db, authContext.organizationId);
-    const contractRepo = new ContractRepository(logger, opts.db, authContext.organizationId);
-    const graphCompositionRepo = new GraphCompositionRepository(logger, opts.db);
     const namespaceRepository = new NamespaceRepository(opts.db, authContext.organizationId);
     const organizationRepo = new OrganizationRepository(logger, opts.db);
-    const composer = new Composer(
-      logger,
-      opts.db,
-      fedGraphRepo,
-      subgraphRepo,
-      contractRepo,
-      graphCompositionRepo,
-      opts.chClient!,
-    );
 
     const cacheWarmerFeature = await organizationRepo.getFeature({
       organizationId: authContext.organizationId,
@@ -89,10 +77,21 @@ export function computeCacheWarmerOperations(
       };
     }
 
-    await composer.fetchAndUploadCacheWarmerOperations({
+    if (!opts.chClient) {
+      return {
+        response: {
+          code: EnumStatusCode.ERR,
+          details: `ClickHouse client is not available`,
+        },
+      };
+    }
+
+    const cacheWarmerRepo = new CacheWarmerRepository(opts.chClient, opts.db);
+    await cacheWarmerRepo.fetchAndUploadCacheWarmerOperations({
       blobStorage: opts.blobStorage,
       federatedGraphId: federatedGraph.id,
       organizationId: authContext.organizationId,
+      logger,
     });
 
     return {
